Minor corrections to make casts explicit and to change the type on rehash to match the rest of the code
git-svn-id: svn://innoue/tracking/trunk@289 4cb6fdf0-3dc9-4801-8ffc-c6659000aa44

+++ common_src/cxx/hash_shelf.cpp
@@ -127,7 +127,7 @@ void Hash_shelf::priv_init()
   
   for(int j = 0; j<rank;++j)
   {
-    hash_dims_[j] = ceil(img_dims_[j]/upb_);
+    hash_dims_[j] = (int)ceil(img_dims_[j]/upb_);
     
     // (((int)(img_dims_[j]))%upb_==0?
     //  ((int)img_dims_[j])/upb_:
@@ -166,7 +166,7 @@ void Hash_shelf::priv_init()
      @param UPB
      new pixel per box value
    */
-void Hash_shelf::rehash(unsigned int UPB){
+void Hash_shelf::rehash(float UPB){
   list<particle*> tmp;
   shelf_to_list(tmp);
   upb_ = UPB;
