+++ common_src/include/hash_shelf.h
@@ -78,7 +78,7 @@ public:
 
 
   ///Changed the pixels per box and re-hashes the shelf
-  void rehash(unsigned int unit_per_box);
+  void rehash(float unit_per_box);
   
   ///Generates the hash value based on a pointer to a particle object
   virtual unsigned int hash_function(const particle* p) const;
